🔧 Fix: 크레딧 충전 컴포넌트 수정
크레딧 충전 컴포넌트 애니메이션 수정 Resolves: #35
