+++ .vscode/settings.json
@@ -10,5 +10,6 @@
 	},
 	"[json]": {
 		"editor.defaultFormatter": "biomejs.biome"
-	}
+	},
+	"cSpell.words": ["Testpage"]
 }
